added userId to localStorage, changed Login and Register Modals as separated components
